requirements.txt
python-docx
langchain==0.2.3
langchain-community 
langchain-core
openai
openpyxl
